requirements.txt
aiohttp==3.11.16
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
click==8.1.8
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
requests==2.32.3
Werkzeug==3.1.3
pandas==1.5.3
numpy==1.24.3
flask-cors==3.0.10
pytesseract==0.3.10
discord.py==2.3.2
marshmallow==3.20.1
pytest-Flask==1.2.0
